requirements.txt
flask==2.3.3
opencv-python==4.8.0.76
numpy==1.24.3
pillow==10.0.0
flask-cors==4.0.0
pyjwt==2.8.0
python-dotenv==1.0.0
